+++ src/core/modules.py
@@ -84,46 +84,46 @@ modules["hash"] = """#
 #########################################################################
 namespace hash
 	func blake2b ($value)
-		py 'self.mem = hashlib.blake2b("' + $value + '".encode()).hexdigest()'
+		python 'self.mem = hashlib.blake2b("' + $value + '".encode()).hexdigest()'
 	endfunc
 	func blake2s ($value)
-		py 'self.mem = hashlib.blake2s("' + $value + '".encode()).hexdigest()'
+		python 'self.mem = hashlib.blake2s("' + $value + '".encode()).hexdigest()'
 	endfunc
 	func md5 ($value)
-		py 'self.mem = hashlib.md5("' + $value + '".encode()).hexdigest()'
+		python 'self.mem = hashlib.md5("' + $value + '".encode()).hexdigest()'
 	endfunc
 	func sha1 ($value)
-		py 'self.mem = hashlib.sha1("' + $value + '".encode()).hexdigest()'
+		python 'self.mem = hashlib.sha1("' + $value + '".encode()).hexdigest()'
 	endfunc
 	func sha224 ($value)
-		py 'self.mem = hashlib.sha224("' + $value + '".encode()).hexdigest()'
+		python 'self.mem = hashlib.sha224("' + $value + '".encode()).hexdigest()'
 	endfunc
 	func sha256 ($value)
-		py 'self.mem = hashlib.sha256("' + $value + '".encode()).hexdigest()'
+		python 'self.mem = hashlib.sha256("' + $value + '".encode()).hexdigest()'
 	endfunc
 	func sha384 ($value)
-		py 'self.mem = hashlib.sha384("' + $value + '".encode()).hexdigest()'
+		python 'self.mem = hashlib.sha384("' + $value + '".encode()).hexdigest()'
 	endfunc
 	func sha3_224 ($value)
-		py 'self.mem = hashlib.sha3_224("' + $value + '".encode()).hexdigest()'
+		python 'self.mem = hashlib.sha3_224("' + $value + '".encode()).hexdigest()'
 	endfunc
 	func sha3_256 ($value)
-		py 'self.mem = hashlib.sha3_256("' + $value + '".encode()).hexdigest()'
+		python 'self.mem = hashlib.sha3_256("' + $value + '".encode()).hexdigest()'
 	endfunc
 	func sha3_384 ($value)
-		py 'self.mem = hashlib.sha3_384("' + $value + '".encode()).hexdigest()'
+		python 'self.mem = hashlib.sha3_384("' + $value + '".encode()).hexdigest()'
 	endfunc
 	func sha3_512 ($value)
-		py 'self.mem = hashlib.sha3_512("' + $value + '".encode()).hexdigest()'
+		python 'self.mem = hashlib.sha3_512("' + $value + '".encode()).hexdigest()'
 	endfunc
 	func sha512 ($value)
-		py 'self.mem = hashlib.sha512("' + $value + '".encode()).hexdigest()'
+		python 'self.mem = hashlib.sha512("' + $value + '".encode()).hexdigest()'
 	endfunc
 	func shake_128 ($value)
-		py 'self.mem = hashlib.shake_128("' + str($value[0]) + '".encode()).hexdigest(' + str($value[1]) + ')'
+		python 'self.mem = hashlib.shake_128("' + str($value[0]) + '".encode()).hexdigest(' + str($value[1]) + ')'
 	endfunc
 	func shake_256 ($value)
-		py 'self.mem = hashlib.shake_256("' + str($value[0]) + '".encode()).hexdigest(' + str($value[1]) + ')'
+		python 'self.mem = hashlib.shake_256("' + str($value[0]) + '".encode()).hexdigest(' + str($value[1]) + ')'
 	endfunc
 endns
 """
@@ -150,10 +150,10 @@ modules["random"] = """#
 #########################################################################
 namespace random
     func randint ($args)
-        py 'self.mem = random.randint(' + str($args[0]) + ',' + str($args[1]) + ')'
+        python 'self.mem = random.randint(' + str($args[0]) + ',' + str($args[1]) + ')'
     endfunc
     func random
-        py 'self.mem = random.random()'
+        python 'self.mem = random.random()'
     endfunc
 endns
 """
@@ -190,15 +190,6 @@ func exit ($code)
     section stdlib_exit_default_exit_with_zero_code
     return 0
 endfunc
-func py
-    python ^
-endfunc
-func sys
-    system ^
-endfunc
-func std_chdir
-    chdir ^
-endfunc
 func std_eval
     eval ^
 endfunc
@@ -206,7 +197,7 @@ func endns
     endnamespace
 endfunc
 func raise ($exdata)
-	mem "self.raise_error('" + $exdata[0] + "', '" + $exdata[1] + "', op)"
+	mem "self.raise_error('" + $exdata[0] + "', '" + $exdata[1] + "', self.operations[self.current_step])"
     python ^
 endfunc
 func assert
@@ -224,8 +215,23 @@ endfunc
 func printl ($value)
     println $value
 endfunc
-struct Object
-endstruct
+func cwd
+    mem os.getcwd()
+endfunc
+func chdir ($path)
+    mem os.chdir($path)
+endfunc
+func typeof ($obj)
+    mem type($obj)
+endfunc
+func system ($cmd)
+    mem os.system($cmd)
+endfunc
+func python
+    rmem exec(^)
+endfunc
+class Object
+endclass
 """
 modules["sys"] = """#
 # sys.pashm
@@ -336,19 +342,19 @@ modules["time"] = """#
 #########################################################################
 namespace time
     func time
-        py 'self.mem = time.time()'
+        python 'self.mem = time.time()'
     endfunc
     func sleep ($time_to_sleep)
-        py 'self.mem = time.sleep(' + str($time_to_sleep) + ')'
+        python 'self.mem = time.sleep(' + str($time_to_sleep) + ')'
     endfunc
     func ctime
-        py 'self.mem = time.ctime()'
+        python 'self.mem = time.ctime()'
     endfunc
     func gmtime
-        py 'self.mem = time.gmtime()'
+        python 'self.mem = time.gmtime()'
     endfunc
     func localtime
-        py 'self.mem = time.localtime()'
+        python 'self.mem = time.localtime()'
     endfunc
 endnamespace
 """