+++ src/operations/classop.py
@@ -1,5 +1,5 @@
 #
-# struct.py
+# classop.py
 #
 # The Pashmak Project
 # Copyright 2020 parsa shahmaleki <parsampsh@gmail.com>
@@ -20,14 +20,14 @@
 # along with Pashmak.  If not, see <https://www.gnu.org/licenses/>.
 #########################################################################
 
-""" Declares a struct """
+""" Declares a class """
 
 import copy
-from core.struct import Struct
+from core.class_system import Class
 
 def run(self, op: dict):
-    """ Declares a struct """
-    self.require_one_argument(op, 'struct operation requires struct name argument')
+    """ Declares a class """
+    self.require_one_argument(op, 'class operation requires class name argument')
     arg = op['args_str']
     arg = arg.split('<', 1)
     parent = None
@@ -37,53 +37,53 @@ def run(self, op: dict):
 
     if '.' in arg:
         self.raise_error(
-            'StructNameContainsDotError', 'name "' + arg + '" for struct contains `.` character', op
+            'ClassNameContainsDotError', 'name "' + arg + '" for class contains `.` character', op
         )
 
     # check parent exists
     parent_real_name = None
     if parent != None:
         try:
-            parent_obj = self.structs[self.current_namespace() + parent]
+            parent_obj = self.classes[self.current_namespace() + parent]
             parent_real_name = self.current_namespace() + parent
         except KeyError:
             parent_obj = None
             for used_namespace in self.used_namespaces:
                 try:
-                    parent_obj = self.structs[used_namespace + '.' + parent]
+                    parent_obj = self.classes[used_namespace + '.' + parent]
                     parent_real_name = used_namespace + '.' + parent
                 except KeyError:
                     pass
             if not parent_obj:
                 try:
-                    parent_obj = self.structs[parent]
+                    parent_obj = self.classes[parent]
                     parent_real_name = parent
                 except KeyError:
-                    self.raise_error('StructError', 'undefined struct "' + parent + '"', op)
+                    self.raise_error('ClassError', 'undefined class "' + parent + '"', op)
                     return
 
-    # check struct already declared
+    # check class already declared
     try:
-        self.structs[self.current_namespace() + arg]
+        self.classes[self.current_namespace() + arg]
         self.raise_error(
-            'StructError',
-            'struct "' + self.current_namespace() + arg + '" already declared',
+            'ClassError',
+            'class "' + self.current_namespace() + arg + '" already declared',
             op
         )
     except KeyError:
         pass
 
     if parent_real_name != None:
-        self.structs[self.current_namespace() + arg] = copy.deepcopy(self.structs[parent_real_name])
-        self.structs[self.current_namespace() + arg].props['__parent__'] = parent_real_name
-        self.structs[self.current_namespace() + arg].props['__name__'] = self.current_namespace() + arg
+        self.classes[self.current_namespace() + arg] = copy.deepcopy(self.classes[parent_real_name])
+        self.classes[self.current_namespace() + arg].props['__parent__'] = parent_real_name
+        self.classes[self.current_namespace() + arg].props['__name__'] = self.current_namespace() + arg
     else:
         if self.current_namespace() + arg != 'Object':
-            self.structs[self.current_namespace() + arg] = copy.deepcopy(self.structs['Object'])
-            self.structs[self.current_namespace() + arg].props['__parent__'] = 'Object'
-            self.structs[self.current_namespace() + arg].props['__name__'] = self.current_namespace() + arg
+            self.classes[self.current_namespace() + arg] = copy.deepcopy(self.classes['Object'])
+            self.classes[self.current_namespace() + arg].props['__parent__'] = 'Object'
+            self.classes[self.current_namespace() + arg].props['__name__'] = self.current_namespace() + arg
         else:
-            self.structs[self.current_namespace() + arg] = copy.deepcopy(Struct(self.current_namespace() + arg, {}))
-            self.structs[self.current_namespace() + arg].props['__parent__'] = None
-            self.structs[self.current_namespace() + arg].props['__name__'] = 'Object'
-    self.current_struct = self.current_namespace() + arg
+            self.classes[self.current_namespace() + arg] = copy.deepcopy(Class(self.current_namespace() + arg, {}))
+            self.classes[self.current_namespace() + arg].props['__parent__'] = None
+            self.classes[self.current_namespace() + arg].props['__name__'] = 'Object'
+    self.current_class = self.current_namespace() + arg