+++ CHANGELOG.md
@@ -1,14 +1,15 @@
 # Pashmak Changelog
 
-## 0.6, 0.6-alpha1 (2020-12-11)
+## 0.6, 0.6-alpha2 (2020-12-11)
 
 - removed `loop` and `while` functions
-- created **Structs** system
+- created **Class** system
 - change default `RuntimeError` name to the real python exception name in exception system
 - created consts system
 - some optimizations in mem eval parser
 - added a feature to use mem symbol `^` in eval, now you can run `my_function ^ + ' something else'`
 - added syntax feature to put `()` in function call syntax, for example `println("hello world")` or `my_func(1, 2)`
+- some optimizations in interpreter code
 
 ## 0.5.4 (2020-12-6)